efarm-backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.16
pluggy==1.2.0
py==1.11.0
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.0.9
pytest==7.1.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.0.4
